Issue #11163: Enforce file size for OverloadMethodsDeclarationOrder input files
