+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/overloadmethodsdeclarationorder/InputOverloadMethodsDeclarationOrder1.java
@@ -6,7 +6,7 @@ OverloadMethodsDeclarationOrder
 
 package com.puppycrawl.tools.checkstyle.checks.coding.overloadmethodsdeclarationorder;
 
-class InputOverloadMethodsDeclarationOrder
+class InputOverloadMethodsDeclarationOrder1
 {
     public void overloadMethod(int i)
     {
@@ -35,7 +35,7 @@ class InputOverloadMethodsDeclarationOrder
         //some foo code
     }
 
-    InputOverloadMethodsDeclarationOrder anonymous = new InputOverloadMethodsDeclarationOrder()
+    InputOverloadMethodsDeclarationOrder1 anonymous = new InputOverloadMethodsDeclarationOrder1()
     {
         public void overloadMethod(int i)
         {
@@ -77,7 +77,7 @@ class InputOverloadMethodsDeclarationOrder
     };
 }
 
-interface Fooable
+interface Fooable1
 {
     public abstract void foo(int i);
     public abstract void foo(String s);
@@ -85,94 +85,4 @@ interface Fooable
 
     // violation below 'All overloaded methods should be placed next to each other.'
     public abstract void foo(String s, Boolean b, int i);
-}
-
-enum FooType {
-    Strategy(""),
-    Shooter(""),
-    RPG("");
-
-    private String description;
-
-    private FooType(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public void overloadMethod(int i)
-    {
-        //some foo code
-    }
-
-    public void overloadMethod(String s)
-    {
-        //some foo code
-    }
-
-    // comments between overloaded methods are allowed.
-    public void overloadMethod(boolean b)
-    {
-        //some foo code
-    }
-
-    public void fooMethod()
-    {
-
-    }
-
-    // violation : because overloads never split
-    // violation below 'All overloaded methods should be placed next to each other.'
-    public void overloadMethod(String s, Boolean b, int i)
-    {
-        //some foo code
-    }
-
-    void test() {}
-
-    String str;
-
-    private interface Testing {}
-
-    // violation below 'All overloaded methods should be placed next to each other.'
-    void test(int x) {}
-
-    private class Inner {
-        void test() {}
-
-        void test(String str) {}
-
-        void test2() {}
-
-        String str;
-
-        // violation below 'All overloaded methods should be placed next to each other.'
-        void test(int x) {}
-    }
-    // violation below 'All overloaded methods should be placed next to each other.'
-    void test(double d) {}
-}
-
-enum Foo2 {
-    VALUE {
-        public void value() {
-            value("");
-        }
-
-        public void middle() {
-        }
-
-        public void value(String s) {
-        }
-    };
-}
-
-@interface ClassPreamble {
-    String author();
 }
